+++ worker/jobs/pull/http.py
@@ -1,4 +1,4 @@
-import requests
+from .base import PullSession
 
 
 def pull_http(source: dict, sink: str) -> str:
@@ -7,12 +7,6 @@ def pull_http(source: dict, sink: str) -> str:
     """
     assert "url" in source, "HTTP source must have a URL"
 
-    session = requests.Session()
-    session.max_redirects = 5
-
-    with session.get(source["url"], stream=True, allow_redirects=True) as response:
-        with open(sink, "wb") as file:
-            for chunk in response:
-                file.write(chunk)
-
+    session = PullSession()
+    session.pull(source["url"], sink)
     return sink