+++ worker/worker.py
@@ -2,13 +2,15 @@
 Stencila Hub Worker.
 
 A Celery app for running jobs.
-See https://docs.celeryproject.org/en/latest/userguide/tasks.html.
 """
 
 import os
 
 from celery import Celery
 
+from jobs.convert import Convert
+from jobs.decode import Decode
+from jobs.encode import Encode
 from jobs.pull import Pull
 from jobs.sleep import Sleep
 
@@ -28,5 +30,9 @@ app.conf.update(
     }
 )
 
+# Register the jobs
+app.register_task(Convert())
+app.register_task(Decode())
+app.register_task(Encode())
 app.register_task(Pull())
 app.register_task(Sleep())